GalxGPT/GalxGPT/web_ui/requirements.txt
Flask==2.1.1
Werkzeug==2.0.3
pandas==1.4.2
openai==1.29.0
gunicorn==20.1.0
python-dotenv==0.19.2
python-docx==0.8.11
pypandoc==1.6.4
openpyxl==3.0.9 
Flask-Session==0.4.0
beautifulsoup4==4.9.3
googlesearch-python==1.1.0
psycopg2-binary==2.9.6
requests==2.25.1
duckduckgo_search==6.2.11b1